+++ src/Factory/SwooleInstalledFactory.php
@@ -12,6 +12,7 @@ class SwooleInstalledFactory
 
     public function __invoke(): callable
     {
+        // TODO regular swoole support is deprecated. Stop checking for it once Shlink v4.0.0 is released
         return static fn (): bool => extension_loaded('openswoole') || extension_loaded('swoole');
     }
 }